safari-534.53-branch/LayoutTests/fast/js/mozilla/strict/15.5.5.2-expected.txt

FAIL 'use strict'; "foo"[0] = 1 should throw an instance of TypeError
PASS "foo"[0] = 1 is 1
PASS true === true
PASS 'use strict'; delete "foo"[0] threw exception of type TypeError.
PASS delete "foo"[0] is false
PASS true === true
 PASSED! 
PASS successfullyParsed is true

TEST COMPLETE
